requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
google-generativeai==0.3.2
pydub==0.25.1
SpeechRecognition
gTTS==2.4.0
reportlab==4.0.7
python-dotenv==1.0.0
google-genai
